+++ crates/runtime/src/vm.rs
@@ -244,22 +244,6 @@ impl KotoVm {
         &self.context.settings.stderr
     }
 
-    /// Returns the named value from the exports map, or None if no matching value is found
-    pub fn get_exported_value(&self, id: &str) -> Option<KValue> {
-        self.exports.data().get(id).cloned()
-    }
-
-    /// Returns the named function from the exports map
-    ///
-    /// None is returned if no matching value is found, or if a matching value is found which isn't
-    /// a callable function.
-    pub fn get_exported_function(&self, id: &str) -> Option<KValue> {
-        match self.get_exported_value(id) {
-            Some(function) if function.is_callable() => Some(function),
-            _ => None,
-        }
-    }
-
     /// Runs the provided [Chunk], returning the resulting [KValue]
     pub fn run(&mut self, chunk: Ptr<Chunk>) -> Result<KValue> {
         // Set up an execution frame to run the chunk in
@@ -297,19 +281,23 @@ impl KotoVm {
         }
     }
 
-    /// Runs a function with some given arguments
-    pub fn run_function(&mut self, function: KValue, args: CallArgs) -> Result<KValue> {
-        self.call_and_run_function(None, function, args)
+    /// Calls a function with some given arguments
+    pub fn call_function<'a>(
+        &mut self,
+        function: KValue,
+        args: impl Into<CallArgs<'a>>,
+    ) -> Result<KValue> {
+        self.call_and_run_function(None, function, args.into())
     }
 
     /// Runs an instance function with some given arguments
-    pub fn run_instance_function(
+    pub fn call_instance_function<'a>(
         &mut self,
         instance: KValue,
         function: KValue,
-        args: CallArgs,
+        args: impl Into<CallArgs<'a>>,
     ) -> Result<KValue> {
-        self.call_and_run_function(Some(instance), function, args)
+        self.call_and_run_function(Some(instance), function, args.into())
     }
 
     fn call_and_run_function(
@@ -328,7 +316,6 @@ impl KotoVm {
         self.registers.push(KValue::Null); // result register
         self.registers.push(instance.unwrap_or_default()); // frame base
         let (arg_count, temp_tuple_values) = match args {
-            CallArgs::None => (0, None),
             CallArgs::Single(arg) => {
                 self.registers.push(arg);
                 (1, None)
@@ -621,10 +608,10 @@ impl KotoVm {
 
                     if let Some(pre_test) = &pre_test {
                         if pre_test.is_callable() {
-                            let pre_test_result = self.run_instance_function(
+                            let pre_test_result = self.call_instance_function(
                                 self_arg.clone(),
                                 pre_test.clone(),
-                                CallArgs::None,
+                                &[],
                             );
 
                             if let Err(error) = pre_test_result {
@@ -633,8 +620,7 @@ impl KotoVm {
                         }
                     }
 
-                    let test_result =
-                        self.run_instance_function(self_arg.clone(), test, CallArgs::None);
+                    let test_result = self.call_instance_function(self_arg.clone(), test, &[]);
 
                     if let Err(error) = test_result {
                         return make_test_error(error, "Error while running test");
@@ -642,10 +628,10 @@ impl KotoVm {
 
                     if let Some(post_test) = &post_test {
                         if post_test.is_callable() {
-                            let post_test_result = self.run_instance_function(
+                            let post_test_result = self.call_instance_function(
                                 self_arg.clone(),
                                 post_test.clone(),
-                                CallArgs::None,
+                                &[],
                             );
 
                             if let Err(error) = post_test_result {
@@ -953,10 +939,8 @@ impl KotoVm {
 
         let non_local = self
             .exports
-            .data()
             .get(name)
-            .cloned()
-            .or_else(|| self.context.prelude.data().get(name).cloned());
+            .or_else(|| self.context.prelude.get(name));
 
         if let Some(non_local) = non_local {
             self.set_register(register, non_local);
@@ -1959,7 +1943,7 @@ impl KotoVm {
         }
 
         for (key_a, value_a) in map_a.data().iter() {
-            let Some(value_b) = map_b.data().get(key_a).cloned() else {
+            let Some(value_b) = map_b.get(key_a) else {
                 return Ok(false);
             };
             match self.run_binary_op(BinaryOp::Equal, value_a.clone(), value_b)? {
@@ -2097,14 +2081,14 @@ impl KotoVm {
         };
 
         // Is the import in the exports?
-        let maybe_in_exports = self.exports.data().get(&import_name).cloned();
+        let maybe_in_exports = self.exports.get(&import_name);
         if let Some(value) = maybe_in_exports {
             self.set_register(import_register, value);
             return Ok(());
         }
 
         // Is the import in the prelude?
-        let maybe_in_prelude = self.context.prelude.data().get(&import_name).cloned();
+        let maybe_in_prelude = self.context.prelude.get(&import_name);
         if let Some(value) = maybe_in_prelude {
             self.set_register(import_register, value);
             return Ok(());
@@ -2184,7 +2168,7 @@ impl KotoVm {
                 let maybe_main = self.exports.get_meta_value(&MetaKey::Main);
                 match maybe_main {
                     Some(main) if main.is_callable() => {
-                        self.run_function(main, CallArgs::None)?;
+                        self.call_function(main, &[])?;
                     }
                     Some(unexpected) => return type_error("callable function", &unexpected),
                     None => {}
@@ -2475,7 +2459,7 @@ impl KotoVm {
                 let mut lookup_map = map.clone();
                 let mut access_result = None;
                 while access_result.is_none() {
-                    let maybe_value = lookup_map.data().get(&key).cloned();
+                    let maybe_value = lookup_map.get(&key);
                     match maybe_value {
                         Some(value) => access_result = Some(value),
                         // Fallback to the map module when there's no metamap
@@ -2526,7 +2510,7 @@ impl KotoVm {
 
                 let mut result = None;
                 if let Some(entries) = o.entries() {
-                    result = entries.data().get(&key).cloned();
+                    result = entries.get(&key);
                 }
 
                 // Iterator fallback?
@@ -2558,8 +2542,8 @@ impl KotoVm {
         iterator_fallback: bool,
         module_name: &str,
     ) -> Result<KValue> {
-        let maybe_op = match module.data().get(key).cloned() {
-            None if iterator_fallback => self.context.core_lib.iterator.data().get(key).cloned(),
+        let maybe_op = match module.get(key) {
+            None if iterator_fallback => self.context.core_lib.iterator.get(key),
             maybe_op => maybe_op,
         };
 
@@ -2679,7 +2663,7 @@ impl KotoVm {
         Ok(())
     }
 
-    fn call_function(
+    fn call_koto_function(
         &mut self,
         call_info: &CallInfo,
         f: &KFunction,
@@ -2751,9 +2735,9 @@ impl KotoVm {
         use KValue::*;
 
         match callable {
-            Function(f) => self.call_function(info, &f, None, temp_tuple_values),
+            Function(f) => self.call_koto_function(info, &f, None, temp_tuple_values),
             CaptureFunction(f) => {
-                self.call_function(info, &f.info, Some(&f.captures), temp_tuple_values)
+                self.call_koto_function(info, &f.info, Some(&f.captures), temp_tuple_values)
             }
             NativeFunction(f) => self.call_external(info, ExternalCallable::Function(f)),
             Object(o) => self.call_external(info, ExternalCallable::Object(o)),
@@ -3182,13 +3166,11 @@ pub(crate) fn clone_generator_vm(vm: &KotoVm) -> Result<KotoVm> {
 
 /// Function call arguments
 ///
-/// This enum provides flexibility in how you'd like to pass arguments to a function.
+/// Typical use will be to use the `From` implementations, either providing a single value that
+/// implements `Into<KValue>`, or an array or slice of `KValue`s.
+///
+/// See [KotoVm::call_function].
 pub enum CallArgs<'a> {
-    /// Indicates that the function will be called without any arguments.
-    ///
-    /// This is used for functions that do not require input from the caller.
-    None,
-
     /// Represents a function call with a single argument.
     Single(KValue),
 
@@ -3197,9 +3179,30 @@ pub enum CallArgs<'a> {
 
     /// Arguments are bundled together as a tuple and then passed to the function.
     ///
-    /// If the function unpacks the tuple in its arguments list then a temporary tuple will be used,
-    /// which avoids the creation of an allocated tuple.
+    /// If the called function unpacks the tuple in its arguments list,
+    /// then a temporary tuple will be used, which avoids the allocation of a regular KTuple.
     AsTuple(&'a [KValue]),
+}
+
+impl<T> From<T> for CallArgs<'static>
+where
+    T: Into<KValue>,
+{
+    fn from(value: T) -> Self {
+        CallArgs::Single(value.into())
+    }
+}
+
+impl<'a> From<&'a [KValue]> for CallArgs<'a> {
+    fn from(args: &'a [KValue]) -> Self {
+        CallArgs::Separate(args)
+    }
+}
+
+impl<'a, const N: usize> From<&'a [KValue; N]> for CallArgs<'a> {
+    fn from(args: &'a [KValue; N]) -> Self {
+        CallArgs::Separate(args.as_ref())
+    }
 }
 
 // A cache of the export maps of imported modules